major refactor of logic layer handlers
