+++ tesseract-server/src/handlers/logic_layer/mod.rs
@@ -1,7 +1,9 @@
 mod aggregate;
 mod detection;
+mod shared;
 
 pub use self::aggregate::ll_aggregate_handler;
 pub use self::aggregate::ll_aggregate_default_handler;
 pub use self::detection::cube_detection_aggregation_handler;
 pub use self::detection::cube_detection_aggregation_default_handler;
+pub use self::shared::{Year, LogicLayerQueryOpt, finish_aggregation};
